requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
discord.py==1.4.1
idna==2.10
isort==4.3.21
jtalkbot==0.5.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pycparser==2.20
pydub==0.24.1
pylint==2.5.3
PyNaCl==1.3.0
pytz==2020.1
six==1.15.0
toml==0.10.1
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
